config.def.h: Set KeePassXC's tags
KeePassXC spawns a client when attempting to perform an auto-type and
more than one entry is matched.

This child client has the same class and instance as the main client, but a
different title. Since the main client changes title, it's not possible to
match it via title.

The child client always has the same title, so it can be matched like
that. But the tags are already set because it has the same class as the
parent. So, if the tag parameter is set to 0, it gets the same tag as the
parent, instead of the currently selected tags.

The solition is to tag the client with every tag, for convenience.

Also, changed qBittorrent's tags to 6.